retiarii_perf/requirements.txt
absl-py==0.10.0
anykeystore==0.2
apex==0.1
astor==0.8.1
autopep8==1.5.4
cachetools==4.1.1
calmsize==0.1.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0.post1
coverage==5.2.1
cryptacular==1.5.5
decorator==4.4.2
defusedxml==0.6.0
filelock==3.0.12
future==0.18.2
google-auth==1.21.0
google-auth-oauthlib==0.4.1
graphviz==0.14.1
grpcio==1.31.0
hupper==1.10.2
hyperopt==0.1.2
idna==2.10
importlib-metadata==1.7.0
install==1.3.3
joblib==0.16.0
json-tricks==3.15.2
Markdown==3.2.2
MarkupSafe==1.1.1
netifaces==0.10.9
networkx==2.5
nni==1.8
numpy==1.19.1
nvidia-dali-cuda100==0.25.0
oauthlib==3.1.0
packaging==20.4
pandas==1.1.1
PasteDeploy==2.1.0
pbkdf2==1.3
Pillow==7.2.0
pkg-resources==0.0.0
pkginfo==1.5.0.1
plaster==1.0
plaster-pastedeploy==0.7
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pyramid==1.10.4
pyramid-mailer==0.15.1
python-dateutil==2.8.1
python3-openid==3.2.0
PythonWebHDFS==0.2.3
pytorch-lightning==0.9.0
pytorch-memlab==0.2.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
repoze.sendmail==4.4.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sacremoses==0.0.43
schema==0.7.2
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.91
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.19
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tokenizers==0.8.1rc1
toml==0.10.1
torch==1.5.1+cu101
torchvision==0.6.1+cu101
tqdm==4.48.2
transaction==3.0.0
transformers==3.0.2
translationstring==1.4
urllib3==1.25.10
velruse==1.1.1
venusian==3.0.0
WebOb==1.8.6
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.1.0
zope.sqlalchemy==1.3
